Add JRuby compatibility by working around JRuby bug
As of JRuby 1.7.6, there's a bug where String#concat(Symbol) will not
raise. This adds a workaround to Vcard for this bug until it's fixed
in JRuby.

Separately, the behavior is odd -- why setting fields and values to
symbols is disallowed is unclear, but given that there's tests and
special code to disallow it, I'll stick with consistency for now.